Always use the referrer used to load the document on same-document navigations
DidCommitProvisionalLoadParams has the `referrer` param that is used
to save the `referrer` attribute in FrameNavigationEntry after commit.

Currently, most same-document navigations will use the initial referrer
that's used to load the document, even after several navigations. An
exception to this is if the navigation is classified as a client-side
redirect, in which case the previous URL is used. This seems to be
accidental, as the `referrer_` that's tracked in the DocumentLoader is
never actually updated for these navigations, so future same-document
navigations will still use the initial referrer that's used to load the
document.

This CL removes the special behavior for client-side redirect same-doc
navigations, to make the behavior more predictable, making it easier
to calculate the `referrer` param in the browser in the future, and also
allows us to remove the redirect chain in the renderer completely in
crrev.com/c/2933038.

Bug: 1171210, 1131832
Change-Id: If460d5fe08ab5795b3dd4d69b0188289d8253839