[ResourceTiming]: Extend TAO-match.html test to cover more resource types
Since we have different code paths for loading different types of
resources, this CL adds tests for Timing-Allow-Origin parsing/checking
that exercise more resource types than before.

This test is actually failing right now for Chrome but should pass in
other browsers. Part of the reason to land this now is to see if other
browsers will need fixes too.

Bug: 1218565
Change-Id: Id274ec1fd914f62d3b805d4d69f81f2945a8baaa